Fixes:
- Convert topicSize and retentionMs from interger to long. So the large topic or long rention topic wont cause sensor parsing error.
- Attach broker id into metrics datapoints to help the engineer find out which brokers are in bad state.

+++ orion-server/src/main/java/com/pinterest/orion/core/automation/sensor/kafka/KafkaTopicSensor.java
@@ -156,7 +156,7 @@ public class KafkaTopicSensor extends KafkaSensor {
         put("clusterId", cluster.getClusterId());
       }};
       // Topic size
-      int topicSize = (int) getTopicSizeByteFromTopicDescription(topicDescription);
+      long topicSize = (long) getTopicSizeByteFromTopicDescription(topicDescription);
       OrionServer.metricsHistogram(
               "kafka.topic.size.bytes",
               topicSize,
@@ -170,7 +170,7 @@ public class KafkaTopicSensor extends KafkaSensor {
               metricsTags
       );
       // Retention
-      int retentionMs= Integer.parseInt(
+      long retentionMs= Long.parseLong(
               topicDescription.getTopicConfigs().getOrDefault("retention.ms", "0.0"));
       OrionServer.metricsHistogram(
               "kafka.topic.retention.ms",
